Fix "missing OAuth parameter" POSTing to auth-url.
The error happened between the GET and the POST to the SP's
authorisation URL. The page served to the GET does not contain any
OAuth-related info; in particular e.g. the request token, which after
this fix is passed as part of the URL still existing in the user's
browser and to which the GET is made.